requirements.txt
# Core dependencies
redis>=4.5.4
pymongo>=4.3.3
requests>=2.31.0
beautifulsoup4>=4.12.2
lxml>=4.9.2

# Web scraping enhanced
fake-useragent>=1.2.1

# Configuration
python-dotenv>=1.0.0

# Logging
loguru>=0.7.0

# Retry mechanism
tenacity>=8.2.2

# Testing
pytest>=7.3.1
